update_phpbb.sh: Only migrate when branches are different.

+++ bin/update_phpbb.sh
@@ -6,6 +6,9 @@ BRANCH=master
 
 cd area51-phpbb3
 git fetch "$REMOTE"
-git reset --hard "$REMOTE/$BRANCH"
-php -f ./phpBB/bin/phpbbcli.php db:migrate
 
+if [ `git rev-parse "$BRANCH"` != `git rev-parse "$REMOTE/$BRANCH"` ]
+then
+	git reset --hard "$REMOTE/$BRANCH"
+	php -f ./phpBB/bin/phpbbcli.php db:migrate
+fi
